[preprocessor/folder] refactor error state log and package
