+++ preprocessor/folder/database_schema.go
@@ -1,7 +1,7 @@
-package preprocessor_folder
+package folder
 
 // schema for file_status table
-var FILE_STATUS_TABLE = `CREATE TABLE IF NOT EXISTS file_status (
+var fileStatusTableSchema = `CREATE TABLE IF NOT EXISTS file_status (
 		id BIGSERIAL PRIMARY KEY,
 		file_path TEXT UNIQUE NOT NULL,
 		sha256 TEXT,
